feat(reimbursements): Show a modal on reimbursement status click
